generalized and moved share model+serializer (ProjectsShared in organization -> ContentShares in climateconnect_api)
