+++ backend/climateconnect_api/models/__init__.py
@@ -21,3 +21,4 @@ from climateconnect_api.models.language import Language
 from climateconnect_api.models.badge import (
     Badge, DonorBadge
 )
+from climateconnect_api.models.content_shares import ContentShares
