+++ backend/climateconnect_api/models/content_shares.py
@@ -4,20 +4,21 @@ from django.contrib.auth.models import User
 
 # Note: Clicking on a specific share button/copying the link or opening a device's native share dialog counts as sharing, 
 # but we don't know whether the user follows through with it.
-class ProjectsShared(models.Model):
+class ContentShares(models.Model):
     project = models.ForeignKey(
         Project,
         related_name="shared_project",
         verbose_name="Project",
         help_text="Points to a project",
-        on_delete=models.CASCADE
+        on_delete=models.CASCADE,
+        null=True
     )
 
     user = models.ForeignKey(
         User,
         related_name="sharing_user",
         verbose_name="Sharing User",
-        help_text="Points to the user who shared the project",
+        help_text="Points to the user who shared the content",
         on_delete=models.CASCADE,
         null=True
     )
@@ -46,22 +47,22 @@ class ProjectsShared(models.Model):
     )
 
     shared_via = models.IntegerField(
-        help_text="Way in which the project was shared",
+        help_text="Way in which the content was shared",
         verbose_name="Shared Via",
         choices=SHARE_OPTIONS,
         default=LINK
     )
 
     created_at = models.DateTimeField(
-        help_text="Time when the user shared the project",
+        help_text="Time when the user shared the content",
         verbose_name="Created At",
         auto_now_add=True
     )
 
     class Meta:
-        app_label = "organization"
-        verbose_name = "Shared Project"
-        verbose_name_plural = "Shared Projects"
+        app_label = "climateconnect_api"
+        verbose_name = "Shared Content"
+        verbose_name_plural = "Shared Content"
         ordering = ["-id"]
 
     def __str__(self):
