+++ backend/organization/models/__init__.py
@@ -24,4 +24,3 @@ from organization.models.translations import (
     ProjectTranslation, OrganizationTranslation,
     PostTranslation, CommentTranslation
 )
-from organization.models.projects_shared import ProjectsShared
